APIs/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-Cors==4.0.1
python-dotenv==1.0.1

psycopg2-binary==2.9.9

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.5
shap==0.45.1
imbalanced-learn==0.12.3

matplotlib==3.9.2
seaborn==0.13.2

requests==2.32.3

ipython==8.25.0
jupyter==1.0.0
